robustify linux quick install script
This makes it so that it will exit early, with a real error message if
you are not running on Ubuntu.

It also moves bluemix cli download from http -> https, and does a
sha256sum on the output to ensure security integrity.

Instead of doing -x, this has a more explicit set of echos for each step.

It also uses a tmpdir to ensure there aren't artifacts left after the
run is over.